pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "page analyzer (project 3 on hexlet)"
authors = ["akasmall <aka_sm@ro.ru>"]
readme = "README.md"
packages = [
  { include = "page_analyzer" },
]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
gunicorn = "^23.0.0"
pylint = "^3.2.7"
psycopg2-binary = "^2.9.9"
jinja2 = "^3.1.4"
requests = "^2.32.3"
validators = "^0.34.0"
beautifulsoup4 = "^4.12.3"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
